requirements.txt
beautifulsoup4==4.12.3
Flask==3.1.0
GitPython==3.1.43
giteapy==1.0.8
PyPDF2==3.0.1
pdfplumber==0.11.4
requests==2.32.3
selenium==4.27.1
seleniumbase==4.33.11
mattermostdriver